Add button onclick handlers; pause till dialog button clicked
Initially, I tried calling ShowDialog() from CancelGame() and NewGame().
However, the code in the calling methods did not wait for the dialog
to close, it ran to completion while waiting for the user to click a
dialog button.

The WaitWhile method in the Utilities class pauses code while waiting
for a condition to change.  In this case, it waits for the DialogResult
to be set before continuing code execution.  This pauses the code while
the dialog is open.