Fix bug in JavaScript timer
If, say, resumeTimer was called multiple times, as happens in game
page, with statuses getting updated multiple times as cells change,
then pauseTimer would need to be clicked multiple times to pause the
timer.

It appeared to be due to calling setTimeout multiple times.  To avoid
that, check if the timer is already running and only call setTimeout if
it isn't.