database: Rename DataAccess model Board to Game
Game represents the starting configuration of a game, with the initial
values, before the player has entered any values of their own.

UserGame is a Game that a user has played or is playing.

GameRun is one attempt at a Game by a player.  The GameRun can
represent a game in progress or a completed game.  A user can have
multiple GameRuns (attempts) for the same Game.